+++ src/ILCompiler.Compiler/src/Compiler/CompilationModuleGroup.cs
@@ -22,17 +22,6 @@ namespace ILCompiler
     public abstract class CompilationModuleGroup
     {
         /// <summary>
-        /// Gets the synthetic assembly that holds types generated by the compiler as part of the compilation process.
-        /// Types and members that declare this module as their owning module are always generated.
-        /// </summary>
-        public ModuleDesc GeneratedAssembly { get; }
-
-        public CompilationModuleGroup(TypeSystemContext context)
-        {
-            GeneratedAssembly = new CompilerGeneratedAssembly(context);
-        }
-
-        /// <summary>
         /// If true, "type" is in the set of input assemblies being compiled
         /// </summary>
         public abstract bool ContainsType(TypeDesc type);
@@ -94,41 +83,5 @@ namespace ILCompiler
         /// accessed through the target platform's import mechanism (ie, Import Address Table on Windows)
         /// </summary>
         public abstract bool CanHaveReferenceThroughImportTable { get; }
-
-        private class CompilerGeneratedAssembly : ModuleDesc, IAssemblyDesc
-        {
-            private MetadataType _globalModuleType;
-
-            public CompilerGeneratedAssembly(TypeSystemContext context)
-                : base(context)
-            {
-                _globalModuleType = new CompilerGeneratedType(this, "<Module>");
-            }
-
-            public override IEnumerable<MetadataType> GetAllTypes()
-            {
-                return Array.Empty<MetadataType>();
-            }
-
-            public override MetadataType GetGlobalModuleType()
-            {
-                return _globalModuleType;
-            }
-
-            public AssemblyName GetName()
-            {
-                return new AssemblyName("System.Private.CompilerGenerated");
-            }
-
-            public override MetadataType GetType(string nameSpace, string name, bool throwIfNotFound = true)
-            {
-                Debug.Fail("Resolving a TypeRef in the compiler generated assembly?");
-
-                if (throwIfNotFound)
-                    ThrowHelper.ThrowTypeLoadException(nameSpace, name, this);
-
-                return null;
-            }
-        }
     }
 }
