+++ src/ILCompiler.Compiler/src/Compiler/SingleFileCompilationModuleGroup.cs
@@ -10,11 +10,6 @@ namespace ILCompiler
 {
     public class SingleFileCompilationModuleGroup : CompilationModuleGroup
     {
-        public SingleFileCompilationModuleGroup(TypeSystemContext context)
-            : base(context)
-        {
-        }
-
         public override bool ContainsType(TypeDesc type)
         {
             return true;
